Add numbers and students ruby files
I was practising earlier before the workshop and that is the time when
I created student script. Then later on in the workshop with Kevin when
I start Ruby basics part 1 and created numbers exercise. Still I am in
the stage to make things work and still not refactoring my code.